pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tristar-backend"
version = "0.1.0"
description = "TriStar — Triangle Smart Targeting and Real-Time Activation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "anthropic>=0.34.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT>=2.8.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "anyio[trio]>=4.4.0",
    "freezegun>=1.5.0",
    "httpx>=0.27.0",  # AsyncClient for testing
    "coverage[toml]>=7.5.0",
    "black>=24.4.0",
    "isort>=5.13.0",
    "ruff>=0.4.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m not integration')",
    "performance: marks tests as performance tests",
]

[tool.coverage.run]
source = ["src/backend"]
omit = [
    "src/backend/core/config.py",  # config-only, no logic
    "tests/*",
]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
